# swagger.yaml
openapi: 3.0.0
paths:
...
  /orders:
    post:
      operationId: OrderController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      security: *ref_4
      summary: Create order from current cart
      tags: *ref_5
    get:
      operationId: OrderController_findAll
      parameters:
        - name: status
          required: true
          in: query
          schema:
            type: string
        - name: userId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: number
                  pageSize:
                    type: number
                  pageNumber:
                    type: number
                  totalPages:
                    type: number
      security: *ref_4
      summary: List all orders (admin)
      tags: *ref_5
  /orders/me:
    get:
      operationId: OrderController_findMyOrders
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: number
                  pageSize:
                    type: number
                  pageNumber:
                    type: number
                  totalPages:
                    type: number
      security: *ref_4
      summary: List current user's orders
      tags: *ref_5
  /orders/{id}/payment-qr:
    get:
      operationId: OrderController_getPaymentQr
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  bankId:
                    type: string
                    example: '970448'
                  accountNumber:
                    type: string
                    example: CASSDVTDAT
                  accountName:
                    type: string
                    example: DOAN VIET TIEN DAT
                  amount:
                    type: number
                    example: 150000
                  memo:
                    type: string
                    example: THANHTOAN 507f1f77bcf86cd799439011
                  qrUrl:
                    type: string
                    example: https://img.vietqr.io/image/970448-CASSDVTDAT-compact2.png?...
      security: *ref_4
      summary: Get VietQR payment details for a pending order
      tags: *ref_5
  /orders/{id}:
    get:
      operationId: OrderController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      security: *ref_4
      summary: Get order by ID
      tags: *ref_5
    delete:
      operationId: OrderController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      security: *ref_4
      summary: Cancel an order
      tags: *ref_5
  /orders/{id}/status:
    patch:
      operationId: OrderController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      security: *ref_4
      summary: Update order status (admin)
      tags: *ref_5
...
  /vouchers:
    post:
      operationId: VoucherController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVoucherDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
      security:
        - bearer: []
      summary: Create voucher (admin)
      tags: &ref_9
        - vouchers
    get:
      operationId: VoucherController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Voucher'
                  total:
                    type: number
                  pageSize:
                    type: number
                  pageNumber:
                    type: number
                  totalPages:
                    type: number
      security:
        - bearer: []
      summary: List all vouchers (admin)
      tags: *ref_9
  /vouchers/validate/{code}:
    get:
      operationId: VoucherController_validateVoucher
      parameters:
        - name: code
          required: true
          in: path
          schema:
            type: string
        - name: subtotal
          required: true
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                  type:
                    type: string
                  value:
                    type: number
                  discountAmount:
                    type: number
                  minOrderAmount:
                    type: number
                    nullable: true
                  maxDiscount:
                    type: number
                    nullable: true
                  validTo:
                    type: string
                    format: date-time
      security:
        - bearer: []
      summary: Validate voucher and preview discount (authenticated)
      tags: *ref_9
  /vouchers/{id}:
    patch:
      operationId: VoucherController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoucherDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voucher'
      security:
        - bearer: []
      summary: Update voucher (admin)
      tags: *ref_9
    delete:
      operationId: VoucherController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
      security:
        - bearer: []
      summary: Soft delete voucher (admin)
      tags: *ref_9
info:
  title: Healthy Food E-Commerce API
  description: Healthy Food E-Commerce API description
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: object
        updatedAt:
          type: object
        deletedAt:
          format: date-time
          type: string
        email:
          type: string
          example: user@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        role:
          type: string
          enum:
            - guest
            - user
            - admin
          example: user
        isActive:
          type: object
          example: true
      required:
        - _id
        - createdAt
        - updatedAt
        - email
        - firstName
        - lastName
        - role
        - isActive
...
    Category:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: object
        updatedAt:
          type: object
        deletedAt:
          format: date-time
          type: string
        name:
          type: string
          example: Vegetables
        slug:
          type: string
          example: vegetables
        description:
          type: string
          example: Fresh organic vegetables
        imageUrl:
          type: string
          example: https://example.com/image.jpg
      required:
        - _id
        - createdAt
        - updatedAt
        - name
        - slug
...
    Product:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: object
        updatedAt:
          type: object
        deletedAt:
          format: date-time
          type: string
        name:
          type: string
          example: Organic Broccoli
        slug:
          type: string
          example: organic-broccoli
        description:
          type: string
          example: Fresh organic broccoli from local farms
        price:
          type: number
          example: 29000
        stock:
          type: number
          example: 100
        imageUrl:
          type: string
          example: https://example.com/broccoli.jpg
        keyCharacteristics:
          example:
            - High protein
            - Omega-3 rich
            - Low GI
          type: array
          items:
            type: string
        weight:
          type: number
          example: 330
        weightUnit:
          type: string
          example: g
        calories:
          type: number
          example: 280
        healthScore:
          type: number
          example: 8.8
        shelfLife:
          type: string
          example: 1 day (refrigerated)
        viewCount:
          type: object
          example: 42
        category:
          $ref: '#/components/schemas/Category'
      required:
        - _id
        - createdAt
        - updatedAt
        - name
        - slug
        - price
        - stock
        - viewCount
        - category
...
    CreateOrderDto:
      type: object
      properties:
        shippingAddress:
          type: string
          example: 123 Nguyen Hue, District 1, HCMC
        deliveryOption:
          type: string
          enum:
            - standard
            - fast
            - urgent
          example: standard
        voucherCode:
          type: string
          example: SUMMER10
        notes:
          type: string
          example: Please call before delivery
      required:
        - shippingAddress
        - deliveryOption
    OrderItem:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: object
        updatedAt:
          type: object
        deletedAt:
          format: date-time
          type: string
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: number
          example: 2
        unitPrice:
          type: number
          example: 29000
      required:
        - _id
        - createdAt
        - updatedAt
        - product
        - quantity
        - unitPrice
    Payment:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: object
        updatedAt:
          type: object
        deletedAt:
          format: date-time
          type: string
        cassoTransactionId:
          type: number
          example: 12345
        amount:
          type: number
          example: 150000
        bankName:
          type: string
          example: Vietcombank
        bankAbbreviation:
          type: string
          example: VCB
        accountNumber:
          type: string
          example: '0123456789'
        description:
          type: string
          example: THANHTOAN 507f1f77bcf86cd799439011
        confirmedAt:
          format: date-time
          type: string
      required:
        - _id
        - createdAt
        - updatedAt
        - cassoTransactionId
        - amount
        - description
        - confirmedAt
    StatusHistoryEntry:
      type: object
      properties:
        status:
          type: string
          example: pending
        changedAt:
          format: date-time
          type: string
        note:
          type: string
      required:
        - status
        - changedAt
    Order:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: object
        updatedAt:
          type: object
        deletedAt:
          format: date-time
          type: string
        user:
          $ref: '#/components/schemas/User'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        payment:
          $ref: '#/components/schemas/Payment'
        status:
          type: string
          enum:
            - pending
            - confirmed
            - shipped
            - delivered
            - cancelled
          example: pending
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/StatusHistoryEntry'
        deliveryOption:
          type: string
          enum:
            - standard
            - fast
            - urgent
          example: standard
        deliveryFee:
          type: number
          example: 20000
        subtotal:
          type: number
          example: 130000
        discountAmount:
          type: number
          example: 0
        voucherCode:
          type: string
          example: SUMMER10
        totalAmount:
          type: number
          example: 150000
        shippingAddress:
          type: string
          example: 123 Nguyen Hue, District 1, HCMC
        notes:
          type: string
          example: Please call before delivery
        trackingNumber:
          type: string
          example: VN123456789
        courierName:
          type: string
          example: Giao Hang Nhanh
        estimatedDeliveryDate:
          format: date-time
          type: string
        actualDeliveryDate:
          format: date-time
          type: string
      required:
        - _id
        - createdAt
        - updatedAt
        - user
        - status
        - statusHistory
        - deliveryOption
        - deliveryFee
        - subtotal
        - discountAmount
        - totalAmount
        - shippingAddress
    UpdateOrderStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - pending
            - confirmed
            - shipped
            - delivered
            - cancelled
          example: shipped
        trackingNumber:
          type: string
          example: VN123456789
        courierName:
          type: string
          example: Giao Hang Nhanh
        estimatedDeliveryDate:
          type: string
          example: '2026-04-05T00:00:00.000Z'
        note:
          type: string
          example: Order confirmed and ready to ship
      required:
        - status
...
    CreateVoucherDto:
      type: object
      properties:
        code:
          type: string
          example: SUMMER10
        type:
          type: string
          enum: &ref_10
            - percent
            - fixed
          example: percent
        value:
          type: number
          example: 10
          description: Percent value or fixed VND amount
        minOrderAmount:
          type: number
          example: 100000
        maxDiscount:
          type: number
          example: 50000
          description: Max discount cap (percent type only)
        usageLimit:
          type: number
          example: 100
        perUserLimit:
          type: number
          example: 1
          description: Max uses per user
        validFrom:
          type: string
          example: '2026-04-15T00:00:00.000Z'
        validTo:
          type: string
          example: '2026-12-31T23:59:59.000Z'
        isActive:
          type: boolean
          example: true
      required:
        - code
        - type
        - value
        - validFrom
        - validTo
    Voucher:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: object
        updatedAt:
          type: object
        deletedAt:
          format: date-time
          type: string
        code:
          type: string
          example: SUMMER10
        type:
          type: string
          enum:
            - percent
            - fixed
          example: percent
        value:
          type: number
          example: 10
          description: Percent (10 = 10%) or fixed amount (VND)
        minOrderAmount:
          type: number
          example: 100000
          description: Min order subtotal to apply voucher
        maxDiscount:
          type: number
          example: 50000
          description: Max discount cap (for percent type)
        usageLimit:
          type: number
          example: 100
          description: Total usage limit across all users
        usedCount:
          type: number
          example: 0
        perUserLimit:
          type: number
          example: 1
          description: Max times a single user can use this voucher
        validFrom:
          format: date-time
          type: string
        validTo:
          format: date-time
          type: string
        isActive:
          type: boolean
          example: true
      required:
        - _id
        - createdAt
        - updatedAt
        - code
        - type
        - value
        - usedCount
        - validFrom
        - validTo
        - isActive
    UpdateVoucherDto:
      type: object
      properties:
        code:
          type: string
          example: SUMMER10
        type:
          type: string
          enum: *ref_10
          example: percent
        value:
          type: number
          example: 10
          description: Percent value or fixed VND amount
        minOrderAmount:
          type: number
          example: 100000
        maxDiscount:
          type: number
          example: 50000
          description: Max discount cap (percent type only)
        usageLimit:
          type: number
          example: 100
        perUserLimit:
          type: number
          example: 1
          description: Max uses per user
        validFrom:
          type: string
          example: '2026-04-15T00:00:00.000Z'
        validTo:
          type: string
          example: '2026-12-31T23:59:59.000Z'
        isActive:
          type: boolean
          example: true